Logo Header, Modal changes, and style updates.
Logo header has been added so now the site can have a logo. This will
only link to the index if it is not on the index. Logo header also has
some links that can be added to a ul similarly to the nav. these do not
display on smaller displays.

Modals now stay at a fixed position on the screen instead of staying at
the top of the screen they should now center in the window.