New plugin, In viewport animation, nav fixes, minor code cleanup.
New free plugin added to check if something is in the viewport. Along
with this I created a function that will make an item animate when it
comes into the viewport.

Nav had an issue where it was not calculating the width correctly due to
the scrollbar appearing after its calculation. The nav is now called
near the end of the document then appended to the body. The headerlogo
also needed to be appended to the body after the nav was to keep it
above.
Some spacing was removed from code.