major headerimg update!
Parallax has been added to the headerimg to add depth to pages.
Javascript has been added to make sure this is not enabled on mobile
devices as parallax support on mobile devices is spotty.
The height of the img has also been greatly increased, and a new test
image has been added to compensate for this.